chore: combined mobile/desktop view for account header
