+++ frontend/src/components/AccountHeaderWidget.vue
@@ -8,14 +8,16 @@
     >
       <template v-slot:text>
         <v-container fluid>
-          <v-row density="compact" class=""
-            ><v-col cols="2"></v-col
-            ><v-col
+          <v-row density="compact" class="">
+            <v-col lg="2" v-if="!smAndDown"></v-col>
+            <v-col
               class="text-primary text-center text-h6 font-weight-bold pa-0 ma-0 ga-0 align-content-center"
-              ><v-card
+            >
+              <v-card
                 class="d-flex align-center justify-center mx-1 px-1 gx-1 bg-secondary-lighten-1"
                 variant="outlined"
-                ><v-tooltip text="Edit Account" location="top">
+              >
+                <v-tooltip text="Edit Account" location="top">
                   <template v-slot:activator="{ props }">
                     <span
                       class="mx-1"
@@ -25,18 +27,20 @@
                       role="button"
                       aria-pressed="false"
                       v-bind="props"
-                      >{{
+                    >
+                      {{
                         account.active
                           ? account.account_name
                           : account.account_name + " (Inactive)"
-                      }}</span
-                    ></template
-                  >
+                      }}
+                    </span>
+                  </template>
                 </v-tooltip>
                 <EditAccountForm
                   v-model="editDialog"
                   :account="account"
-                  @update-dialog="updateEditDialog" />
+                  @update-dialog="updateEditDialog"
+                />
                 <v-tooltip
                   :text="account.active ? 'Delete Account' : 'Enable Account'"
                   location="top"
@@ -58,13 +62,18 @@
                 <DeleteAccountForm
                   v-model="deleteDialog"
                   :account="account"
-                  @update-dialog="updateDeleteDialog" /></v-card></v-col
-            ><v-col cols="2"></v-col></v-row
-          ><v-row density="compact"
-            ><v-col class="text-center align-content-end"
-              ><v-tooltip text="Adjust Balance" location="top">
-                <template v-slot:activator="{ props }"
-                  ><div
+                  @update-dialog="updateDeleteDialog"
+                />
+              </v-card>
+            </v-col>
+            <v-col cols="2" v-if="!smAndDown"></v-col>
+          </v-row>
+          <!-- Large Display View -->
+          <v-row density="compact" v-if="!smAndDown">
+            <v-col class="text-center align-content-end">
+              <v-tooltip text="Adjust Balance" location="top">
+                <template v-slot:activator="{ props }">
+                  <div
                     class="text-accent-lighten-1 font-weight-bold text-h4 d-inline-block"
                     @click="adjBalDialog = true"
                     tabindex="0"
@@ -77,50 +86,56 @@
                     <NumberFlow
                       :value="account.balance"
                       :format="{ style: 'currency', currency: 'USD' }"
-                    /></div></template></v-tooltip
-              ><AdjustBalanceForm
+                    />
+                  </div>
+                </template>
+              </v-tooltip>
+              <AdjustBalanceForm
                 v-model="adjBalDialog"
                 :account="account"
                 @update-dialog="updateAdjBalDialog"
               />
-              <div class="text-secondary-lighten-2">current balance</div></v-col
-            >
+              <div class="text-secondary-lighten-2">current balance</div>
+            </v-col>
             <v-col
               v-if="account.account_type.id == 1"
               class="text-center align-content-end"
-              ><div class="text-white font-weight-bold text-body">
+            >
+              <div class="text-white font-weight-bold text-body">
                 {{
                   account.next_cycle_date
                     ? formatDate(account.next_cycle_date)
                     : "n/a"
                 }}
               </div>
-              <div class="text-secondary-lighten-2">
-                statement end date
-              </div></v-col
-            ><v-col
+              <div class="text-secondary-lighten-2">statement end date</div>
+            </v-col>
+            <v-col
               v-if="account.account_type.id == 1"
               class="text-center align-content-end"
-              ><div class="text-white font-weight-bold text-body">
+            >
+              <div class="text-white font-weight-bold text-body">
                 <NumberFlow
                   :value="account.last_statement_amount"
                   :format="{ style: 'currency', currency: 'USD' }"
                 />
               </div>
-              <div class="text-secondary-lighten-2">
-                last statement balance
-              </div></v-col
-            ><v-col
+              <div class="text-secondary-lighten-2">last statement balance</div>
+            </v-col>
+            <v-col
               v-if="account.account_type.id == 1"
               class="text-center align-content-end"
-              ><div class="text-white font-weight-bold text-body">
+            >
+              <div class="text-white font-weight-bold text-body">
                 {{ account.due_date ? formatDate(account.due_date) : "n/a" }}
               </div>
-              <div class="text-secondary-lighten-2">due date</div></v-col
-            ><v-col
+              <div class="text-secondary-lighten-2">due date</div>
+            </v-col>
+            <v-col
               v-if="account.account_type.id == 1"
               class="text-center align-content-end"
-              ><div
+            >
+              <div
                 class="text-white font-weight-bold text-body d-inline-block"
                 @click="handleClick"
                 tabindex="0"
@@ -133,22 +148,140 @@
                   :format="{ style: 'currency', currency: 'USD' }"
                 />
               </div>
-              <div class="text-secondary-lighten-2">rewards</div></v-col
-            ><v-col
+              <div class="text-secondary-lighten-2">rewards</div>
+            </v-col>
+            <v-col
               v-if="account.account_type.id == 1"
               class="text-center align-content-end"
-              ><div class="text-white font-weight-bold text-body">
+            >
+              <div class="text-white font-weight-bold text-body">
                 <NumberFlow
                   :value="account.available_credit"
                   :format="{ style: 'currency', currency: 'USD' }"
                 />
               </div>
-              <div class="text-secondary-lighten-2">
-                available credit
-              </div></v-col
-            ></v-row
-          >
+              <div class="text-secondary-lighten-2">available credit</div>
+            </v-col>
+          </v-row>
+          <!-- Small Display View -->
+          <v-row density="compact" v-if="smAndDown">
+            <v-col class="text-center align-content-end">
+              <v-tooltip text="Adjust Balance" location="top">
+                <template v-slot:activator="{ props }">
+                  <div
+                    class="text-accent-lighten-1 font-weight-bold text-h4 d-inline-block"
+                    @click="adjBalDialog = true"
+                    tabindex="0"
+                    @keydown.enter="adjBalDialog = true"
+                    role="button"
+                    aria-pressed="false"
+                    v-bind="props"
+                    width="200"
+                  >
+                    <NumberFlow
+                      :value="account.balance"
+                      :format="{ style: 'currency', currency: 'USD' }"
+                    />
+                  </div>
+                </template>
+              </v-tooltip>
+              <AdjustBalanceForm
+                v-model="adjBalDialog"
+                :account="account"
+                @update-dialog="updateAdjBalDialog"
+              />
+              <div class="text-secondary-lighten-2">current balance</div>
+            </v-col>
+          </v-row>
+          <v-row density="compact" v-if="smAndDown">
+            <v-col class="text-center">
+              <v-btn
+                size="x-small"
+                variant="text"
+                :append-icon="!showMore ? 'mdi-chevron-down' : 'mdi-chevron-up'"
+                @click="toggleMore"
+                v-if="account.account_type.id == 1"
+              >
+                {{ !showMore ? "more" : "less" }}
+              </v-btn>
+            </v-col>
+          </v-row>
         </v-container>
+        <v-expand-transition>
+          <v-container fluid v-if="account.account_type.id == 1 && showMore">
+            <v-row density="compact" class="">
+              <v-col
+                v-if="account.account_type.id == 1"
+                class="text-center align-content-end"
+              >
+                <div class="text-white font-weight-bold text-body">
+                  {{
+                    account.next_cycle_date
+                      ? formatDate(account.next_cycle_date)
+                      : "n/a"
+                  }}
+                </div>
+                <div class="text-secondary-lighten-2">statement end date</div>
+              </v-col>
+              <v-col
+                v-if="account.account_type.id == 1"
+                class="text-center align-content-end"
+              >
+                <div class="text-white font-weight-bold text-body">
+                  <NumberFlow
+                    :value="account.last_statement_amount"
+                    :format="{ style: 'currency', currency: 'USD' }"
+                  />
+                </div>
+                <div class="text-secondary-lighten-2">
+                  last statement balance
+                </div>
+              </v-col>
+            </v-row>
+            <v-row density="compact">
+              <v-col
+                v-if="account.account_type.id == 1"
+                class="text-center align-content-end"
+              >
+                <div class="text-white font-weight-bold text-body">
+                  {{ account.due_date ? formatDate(account.due_date) : "n/a" }}
+                </div>
+                <div class="text-secondary-lighten-2">due date</div>
+              </v-col>
+              <v-col
+                v-if="account.account_type.id == 1"
+                class="text-center align-content-end"
+              >
+                <div
+                  class="text-white font-weight-bold text-body d-inline-block"
+                  @click="handleClick"
+                  tabindex="0"
+                  @keydown.enter="handleClick"
+                  role="button"
+                  aria-pressed="false"
+                >
+                  <NumberFlow
+                    :value="account.rewards_amount"
+                    :format="{ style: 'currency', currency: 'USD' }"
+                  />
+                </div>
+                <div class="text-secondary-lighten-2">rewards</div>
+              </v-col>
+              <v-col
+                v-if="account.account_type.id == 1"
+                class="text-center align-content-end"
+              >
+                <div class="text-white font-weight-bold text-body">
+                  <NumberFlow
+                    :value="account.available_credit"
+                    :format="{ style: 'currency', currency: 'USD' }"
+                  />
+                </div>
+                <div class="text-secondary-lighten-2">available credit</div>
+              </v-col>
+            </v-row>
+          </v-container>
+        </v-expand-transition>
       </template>
     </v-card>
     <v-skeleton-loader
@@ -160,39 +293,45 @@
   </div>
 </template>
 <script setup>
-import { defineProps, ref } from "vue";
-import { useAccountByID } from "@/composables/accountsComposable";
-import EditAccountForm from "./EditAccountForm.vue";
-import AdjustBalanceForm from "./AdjustBalanceForm.vue";
-import DeleteAccountForm from "./DeleteAccountForm.vue";
-import NumberFlow from "@number-flow/vue";
+  import { defineProps, ref } from "vue";
+  import { useAccountByID } from "@/composables/accountsComposable";
+  import EditAccountForm from "./EditAccountForm.vue";
+  import AdjustBalanceForm from "./AdjustBalanceForm.vue";
+  import DeleteAccountForm from "./DeleteAccountForm.vue";
+  import NumberFlow from "@number-flow/vue";
+  import { useDisplay } from "vuetify";
 
-const adjBalDialog = ref(false);
-const editDialog = ref(false);
-const deleteDialog = ref(false);
+  const { smAndDown } = useDisplay();
+  const adjBalDialog = ref(false);
+  const editDialog = ref(false);
+  const deleteDialog = ref(false);
+  const showMore = ref(false);
 
-const props = defineProps({
-  account: Array,
-});
+  const props = defineProps({
+    account: Array,
+  });
 
-const { account, isLoading } = useAccountByID(props.account);
+  const { account, isLoading } = useAccountByID(props.account);
 
-const updateAdjBalDialog = value => {
-  adjBalDialog.value = value;
-};
-const updateEditDialog = value => {
-  editDialog.value = value;
-};
-const updateDeleteDialog = value => {
-  deleteDialog.value = value;
-};
-const formatDate = dateString => {
-  const date = new Date(dateString + "T00:00:00Z");
-  return new Intl.DateTimeFormat("en-US", {
-    year: "numeric",
-    month: "long",
-    day: "numeric",
-    timeZone: "UTC",
-  }).format(date);
-};
+  const updateAdjBalDialog = value => {
+    adjBalDialog.value = value;
+  };
+  const updateEditDialog = value => {
+    editDialog.value = value;
+  };
+  const updateDeleteDialog = value => {
+    deleteDialog.value = value;
+  };
+  const formatDate = dateString => {
+    const date = new Date(dateString + "T00:00:00Z");
+    return new Intl.DateTimeFormat("en-US", {
+      year: "numeric",
+      month: "long",
+      day: "numeric",
+      timeZone: "UTC",
+    }).format(date);
+  };
+  const toggleMore = () => {
+    showMore.value = !showMore.value;
+  };
 </script>
